kaspa client refactor
